server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.47
aiosqlite==0.21.0
bcrypt==4.3.0
PyJWT==2.10.1
python-multipart==0.0.20
httpx==0.28.1
pydantic-settings==2.8.1
email-validator==2.2.0
websockets==15.0.1

# dev
pytest==8.3.5
pytest-asyncio==0.25.3
